renames classes and packages in integration tests
