+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koclass/snippet/forkoresideinpackageprovider/class-not-reside-in-file-package.kttxt
