+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koclass/snippet/forkoresideinpackageprovider/class-reside-outside-file-package.kttxt
