+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koconstructor/KoConstructorDeclarationForKoResideInPackageProviderTest.kt
@@ -1,17 +1,17 @@
-package com.lemonappdev.konsist.core.declaration.koenumconstant
+package com.lemonappdev.konsist.core.declaration.koconstructor
 
 import com.lemonappdev.konsist.TestSnippetProvider.getSnippetKoScope
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 
-class KoEnumConstantDeclarationForKoResideInOrOutsidePackageProviderTest {
+class KoConstructorDeclarationForKoResideInPackageProviderTest {
     @Test
-    fun `enum-const-not-reside-in-file-package`() {
+    fun `constructor-not-reside-in-file-package`() {
         // given
-        val sut = getSnippetFile("enum-const-not-reside-in-file-package")
+        val sut = getSnippetFile("constructor-not-reside-in-file-package")
             .classes()
             .first()
-            .enumConstants
+            .constructors
             .first()
 
         // then
@@ -19,12 +19,12 @@ class KoEnumConstantDeclarationForKoResideInOrOutsidePackageProviderTest {
     }
 
     @Test
-    fun `enum-const-reside-in-file-package`() {
+    fun `constructor-reside-in-file-package`() {
         // given
-        val sut = getSnippetFile("enum-const-reside-in-file-package")
+        val sut = getSnippetFile("constructor-reside-in-file-package")
             .classes()
             .first()
-            .enumConstants
+            .constructors
             .first()
 
         // then
@@ -32,12 +32,12 @@ class KoEnumConstantDeclarationForKoResideInOrOutsidePackageProviderTest {
     }
 
     @Test
-    fun `enum-const-not-reside-outside-file-package`() {
+    fun `constructor-not-reside-outside-file-package`() {
         // given
-        val sut = getSnippetFile("enum-const-not-reside-outside-file-package")
+        val sut = getSnippetFile("constructor-not-reside-outside-file-package")
             .classes()
             .first()
-            .enumConstants
+            .constructors
             .first()
 
         // then
@@ -45,12 +45,12 @@ class KoEnumConstantDeclarationForKoResideInOrOutsidePackageProviderTest {
     }
 
     @Test
-    fun `enum-const-reside-outside-file-package`() {
+    fun `constructor-reside-outside-file-package`() {
         // given
-        val sut = getSnippetFile("enum-const-reside-outside-file-package")
+        val sut = getSnippetFile("constructor-reside-outside-file-package")
             .classes()
             .first()
-            .enumConstants
+            .constructors
             .first()
 
         // then
@@ -58,5 +58,5 @@ class KoEnumConstantDeclarationForKoResideInOrOutsidePackageProviderTest {
     }
 
     private fun getSnippetFile(fileName: String) =
-        getSnippetKoScope("core/declaration/koenumconstant/snippet/forkoresideinoroutsidepackageprovider/", fileName)
+        getSnippetKoScope("core/declaration/koconstructor/snippet/forkoresideinpackageprovider/", fileName)
 }
