+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koconstructor/snippet/forkoresideinpackageprovider/constructor-not-reside-in-file-package.kttxt
