+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koconstructor/snippet/forkoresideinpackageprovider/constructor-reside-in-file-package.kttxt
