+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koenumconstant/snippet/forkoresideinpackageprovider/enum-const-not-reside-outside-file-package.kttxt
