+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kofunction/snippet/forkoresideinpackageprovider/function-not-reside-in-file-package.kttxt
