+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kofunction/snippet/forkoresideinpackageprovider/function-reside-in-file-package.kttxt
