+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kointerface/KoInterfaceDeclarationForKoResideInPackageProviderTest.kt
@@ -4,7 +4,7 @@ import com.lemonappdev.konsist.TestSnippetProvider.getSnippetKoScope
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 
-class KoInterfaceDeclarationForKoResideInOrOutsidePackageProviderTest {
+class KoInterfaceDeclarationForKoResideInPackageProviderTest {
     @Test
     fun `interface-not-reside-in-file-package`() {
         // given
@@ -50,5 +50,5 @@ class KoInterfaceDeclarationForKoResideInOrOutsidePackageProviderTest {
     }
 
     private fun getSnippetFile(fileName: String) =
-        getSnippetKoScope("core/declaration/kointerface/snippet/forkoresideinoroutsidepackageprovider/", fileName)
+        getSnippetKoScope("core/declaration/kointerface/snippet/forkoresideinpackageprovider/", fileName)
 }
