+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kointerface/snippet/forkoresideinpackageprovider/interface-not-reside-in-file-package.kttxt
