+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kointerface/snippet/forkoresideinpackageprovider/interface-reside-in-file-package.kttxt
