+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kointerface/snippet/forkoresideinpackageprovider/interface-reside-outside-file-package.kttxt
