+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koobject/snippet/forkoresideinpackageprovider/object-reside-in-file-package.kttxt
