+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparameter/snippet/forkoresideinpackageprovider/parameter-not-reside-in-file-package.kttxt
