+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparameter/snippet/forkoresideinpackageprovider/parameter-reside-in-file-package.kttxt
