+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparameter/snippet/forkoresideinpackageprovider/parameter-reside-outside-file-package.kttxt
