+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparent/snippet/forkoresideinpackageprovider/parent-of-class-not-reside-outside-file-package.kttxt
