+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparent/snippet/forkoresideinpackageprovider/parent-of-class-reside-in-file-package.kttxt
