+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparent/snippet/forkoresideinpackageprovider/parent-of-class-reside-outside-file-package.kttxt
