+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparent/snippet/forkoresideinpackageprovider/parent-of-object-not-reside-outside-file-package.kttxt
