+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparent/snippet/forkoresideinpackageprovider/parent-of-object-reside-in-file-package.kttxt
