+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koproperty/snippet/forkoresideinpackageprovider/property-not-reside-in-file-package.kttxt
