+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koproperty/snippet/forkoresideinpackageprovider/property-reside-outside-file-package.kttxt
