+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kotypealias/snippet/forkoresideinpackageprovider/typealias-not-reside-in-file-package.kttxt
