+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kotypealias/snippet/forkoresideinpackageprovider/typealias-reside-outside-file-package.kttxt
